closed_loop/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.116.0
fonttools==4.58.5
frozenlist==1.7.0
h11==0.16.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
multidict==6.6.3
numpy==2.2.6
opencv-python==4.12.0.88
OpenEXR==3.3.4
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
